@@ -9,9 +9,9 @@
     "@testing-library/user-event": "^12.8.3",
     "bootstrap": "^4.6.0",
     "bootstrap-icons": "^1.6.0",
-    "react": "^17.0.2",
+    "react": "^18.0.0",
     "react-bootstrap": "^1.6.4",
-    "react-dom": "^17.0.2",
+    "react-dom": "^18.0.0",
     "react-router-dom": "^6.2.1",
     "react-scripts": "4.0.3",
     "react-toastify": "^8.1.1",
